# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mosaic API
  version: 1.0.0
  description: |
    Zero-knowledge encrypted photo gallery API.
    
    The server never sees plaintext photos or metadata—all encryption/decryption 
    happens client-side. The server stores encrypted blobs and facilitates 
    synchronization between clients.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: API base path (behind reverse proxy)
security:
  - RemoteUser: []
...
paths:
...
  /shards/{shardId}:
    get:
      operationId: downloadShard
      summary: Download encrypted shard
      tags: [Shards]
      parameters:
        - name: shardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encrypted shard data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /s/{linkIdBase64}:
    get:
      operationId: getShareLinkInfo
      summary: Get share link info (anonymous)
      description: |
        Returns basic info about a share link without authentication.
        The linkIdBase64 is derived from the link secret (first 16 bytes
        of BLAKE2b hash), URL-safe base64 encoded.
      tags: [Anonymous Access]
      security: []
      parameters:
        - name: linkIdBase64
          in: path
          required: true
          schema:
            type: string
          description: URL-safe base64 encoded link ID
      responses:
        '200':
          description: Link info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkAccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Link expired or revoked
  /s/{linkIdBase64}/keys:
    get:
      operationId: getShareLinkKeys
      summary: Get wrapped tier keys for share link (anonymous)
      description: |
        Returns wrapped tier keys for all epochs accessible via this link.
        Keys are encrypted with the link's wrapping key (derived from secret).
      tags: [Anonymous Access]
      security: []
      parameters:
        - name: linkIdBase64
          in: path
          required: true
          schema:
            type: string
          description: URL-safe base64 encoded link ID
      responses:
        '200':
          description: Wrapped epoch keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkEpochKeyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Link expired or revoked
  /s/{linkIdBase64}/photos:
    get:
      operationId: getShareLinkPhotos
      summary: Get photo manifests via share link (anonymous)
      description: |
        Returns encrypted photo manifests accessible via this link.
        Photos are filtered by tier and deletion status.
      tags: [Anonymous Access]
      security: []
      parameters:
        - name: linkIdBase64
          in: path
          required: true
          schema:
            type: string
          description: URL-safe base64 encoded link ID
        - $ref: '#/components/parameters/PageSkip'
        - $ref: '#/components/parameters/PageTake'
      responses:
        '200':
          description: List of photo manifests
          headers:
            X-Pagination-Skip:
              $ref: '#/components/headers/PaginationSkip'
            X-Pagination-Take:
              $ref: '#/components/headers/PaginationTake'
            X-Pagination-Total-Count:
              $ref: '#/components/headers/PaginationTotalCount'
            X-Pagination-Has-More:
              $ref: '#/components/headers/PaginationHasMore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedShareLinkPhotos'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Link expired or revoked
  /s/{linkIdBase64}/shards/{shardId}:
    get:
      operationId: getShareLinkShard
      summary: Download encrypted shard via share link (anonymous)
      description: |
        Downloads an encrypted shard blob via share link without authentication.
        The shard must belong to a manifest in the linked album.
      tags: [Anonymous Access]
      security: []
      parameters:
        - name: linkIdBase64
          in: path
          required: true
          schema:
            type: string
          description: URL-safe base64 encoded link ID
        - name: shardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Shard UUID
      responses:
        '200':
          description: Encrypted shard blob
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Link expired or revoked

  # ===========================================================================
  # Admin - Settings
  # ===========================================================================
...
components:
...
  securitySchemes:
    RemoteUser:
      type: apiKey
      in: header
      name: Remote-User
      description: |
        User identity from trusted reverse proxy.
        Only accepted from configured trusted IP ranges.
    cookieAuth:
      type: apiKey
      in: cookie
      name: mosaic_session
      description: |
        LocalAuth session cookie. The backend sets this cookie as HttpOnly with Path=/api,
        SameSite=Lax for development/testing HTTP, and Secure + SameSite=Strict in production.

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
...
    PageSkip:
      name: skip
      in: query
      required: false
      description: Number of rows to skip before returning a page.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageTake:
      name: take
      in: query
      required: false
      description: Page size. Values are clamped to the backend maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  headers:
    PaginationSkip:
      description: Normalized skip value applied by the backend.
      schema:
        type: integer
    PaginationTake:
      description: Normalized take value applied by the backend.
      schema:
        type: integer
    PaginationTotalCount:
      description: Total rows available for the query before paging.
      schema:
        type: integer
    PaginationHasMore:
      description: Whether another page exists after this response.
      schema:
        type: boolean

  # ===========================================================================
  # Responses
  # ===========================================================================
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 problem details error response.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
          description: Request correlation identifier when available.
      additionalProperties: true
    PagedResult:
      type: object
      required: [items, nextSkip]
      properties:
        items:
          type: array
          items: {}
        nextSkip:
          type: integer
          nullable: true
          description: Skip value for the next page, or null when no further page is available.
...
    PagedShareLinkPhotos:
      allOf:
        - $ref: '#/components/schemas/PagedResult'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ShareLinkPhotoResponse'
...
    LinkAccessResponse:
      type: object
      required:
        - albumId
        - accessTier
        - epochCount
      properties:
        albumId:
          type: string
          format: uuid
        accessTier:
          type: integer
          enum: [1, 2, 3]
        epochCount:
          type: integer
          description: Number of epochs with keys for this link
    LinkEpochKeyResponse:
      type: object
      required:
        - epochId
        - tier
        - nonce
        - encryptedKey
      properties:
        epochId:
          type: integer
        tier:
          type: integer
          enum: [1, 2, 3]
        nonce:
          type: string
          format: byte
          description: Base64-encoded 24-byte nonce
        encryptedKey:
          type: string
          format: byte
          description: Base64-encoded encrypted tier key
        signPubkey:
          type: string
          description: Base64-encoded epoch signing public key
    ShareLinkPhotoResponse:
      type: object
      required:
        - id
        - versionCreated
        - isDeleted
        - encryptedMeta
        - signature
        - signerPubkey
        - shardIds
      properties:
        id:
          type: string
          format: uuid
        versionCreated:
          type: integer
          format: int64
        isDeleted:
          type: boolean
        encryptedMeta:
          type: string
          format: byte
          description: Base64-encoded encrypted photo metadata
        signature:
          type: string
          description: Base64-encoded Ed25519 signature
        signerPubkey:
          type: string
          description: Base64-encoded epoch signing public key
        shardIds:
          type: array
          items:
            type: string
            format: uuid
    # -------------------------------------------------------------------------
    # Admin - Quota Settings
    # -------------------------------------------------------------------------
...